services/api/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pillow
openai>=1.0.0
requests>=2.31.0

# Database migrations
alembic

# Task queue (better than Celery)
rq
redis

# Monitoring and logging
sentry-sdk[fastapi]
structlog
prometheus-client

# Testing
pytest
pytest-asyncio
pytest-cov
httpx
faker

# Code quality
black
isort
flake8
mypy